tests/wpt/tests/fedcm/fedcm-multi-idp/fedcm-multi-idp-abort.https.html

<!DOCTYPE html>
<title>Federated Credential Management API multi IDP abort.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script type="module">
import {request_options_with_two_idps,
        fedcm_test,
        fedcm_get_and_select_first_account} from '../support/fedcm-helper.sub.js';

fedcm_test(async t => {
  let controller = new AbortController();
  let test_options = request_options_with_two_idps();
  test_options.signal = controller.signal;
  const cred = fedcm_get_and_select_first_account(t, test_options);
  controller.abort();
  return promise_rejects_dom(t, 'AbortError', cred);
}, "Test that the abort signal works when multiple IDPs are used.");
</script>
